Refactor directories / imports to implementation files
